refactor: rename TerminalDataEvent package to event
